pi/requirements.txt
fastapi==0.110.0
uvicorn==0.20.0
python-multipart==0.0.6
google-auth
requests
Pillow
Inky
